+++ tests/fake_nclink_server.c
@@ -3,6 +3,7 @@
 
 #include "fake_nclink_server.h"
 
+#include <stdio.h>
 #include <stdlib.h>
 #include <string.h>